Clean fileupload route in dep_check
