+++ microservices/dep_check/routes.py
@@ -1,15 +1,11 @@
-import glob
-import json
-import os
-import shutil
 from pathlib import Path
 
-from flask import Blueprint, Response, flash, request
+from flask import Blueprint, Response, request
 from flask_cors import CORS
 from google.cloud import storage
 
+from depscan import depscan, serialise_request_files
 from fire import authenticateToken, deleteDB, setDB
-from parsers import checkFile
 
 # Registering routes to 'routes' so they can be accessed in other files
 routes = Blueprint("urls", __name__,)
@@ -22,23 +18,6 @@ gcs = storage.Client()
 bucket = gcs.get_bucket("vuldoctor2")
 
 
-def allowed_file(filename):
-    return "." in filename and filename.rsplit(".", 1)[1].lower() in {
-        "py",
-        "java",
-        "js",
-        "json",
-        "xml",
-        "lock",
-        "gradle",
-        "kts",
-        "sbt",
-        "txt",
-        "sum",
-        "csproj",
-    }
-
-
 @routes.route("/", methods=["POST"])
 def upload_file():
 
@@ -46,75 +25,31 @@ def upload_file():
     try:
         decoded_token = authenticateToken(request)
         uid = decoded_token["uid"]
-    except:
+        projectid = request.args.get("projectID")
+    except Exception as e:
+        print(e)
         resp = Response(status=403)
         return resp
 
-    # Serialise files onto disk
-    projectid = request.args.get("projectID")
+    # Set Processing
     setDB(uid, projectid)
+
+    # Create working directory
     filedir = "upload/" + uid + "/" + projectid + "/"
     Path(filedir).mkdir(parents=True, exist_ok=True)
-    files = []
-    if "file" not in request.files:
-        flash("No file part")
+
+    # Serialise files from request
+    files = serialise_request_files(request, filedir)
+    if not files:
         return Response(status=204)
-    n_files = 0
-    n_allowed = 0
-    for file in request.files.getlist("file"):
-        if file.filename == "":
-            flash("No selected file")
-            return Response(status=204)
-        if file:
-            n_files += 1
-        if file and allowed_file(file.filename):
-            file.save(filedir + file.filename)
-            files.append(filedir + file.filename)
-            n_allowed += 1
 
-    # Extracted import data
-    extracted_imports = []
-    for file in files:
-        extracted_imports.append(checkFile(file))
-    # print(extracted_imports);
+    # Perform dep scanning
+    depscan_result = depscan(files, filedir, bucket, uid, projectid)
 
-    # Save extracted dependencies
-    blob = bucket.blob(uid + "/" + projectid + "/" + "extracted_deps.json")
-    blob.upload_from_string(
-        data=json.dumps(extracted_imports), content_type="application/json"
-    )
-
-    # Run dep-scan
-    src = filedir
-    rpt = filedir + "reports/depscan.json"
-    os.system("scan --src {} --report_file {} --sync".format(src, rpt))
-
-    # Get project type and upload results
-    results_files = glob.glob(filedir + "reports/*")
-    if len(results_files) == 0:
+    if not depscan_result:
         deleteDB(uid, projectid)
         return Response(status=400)
 
-    depscan_file = [i for i in results_files if "depscan" in i]
-    if len(depscan_file) == 0:
-        print("No vulnerabilities detected.")
-        bom_file = [i for i in results_files if "bom" in i][0]
-        project_type = bom_file.split("-")[1].split(".")[0]
-        blob = bucket.blob(uid + "/" + projectid + "/" + "bom.json")
-        blob.upload_from_filename(filedir + "reports/bom-{}.json".format(project_type))
-        blob = bucket.blob(uid + "/" + projectid + "/" + "depscan.json")
-        blob.upload_from_string(data=json.dumps({}), content_type="application/json")
-    else:
-        depscan_file = depscan_file[0]
-        project_type = depscan_file.split("-")[1].split(".")[0]
-        blob = bucket.blob(uid + "/" + projectid + "/" + "bom.json")
-        blob.upload_from_filename(filedir + "reports/bom-{}.json".format(project_type))
-        blob = bucket.blob(uid + "/" + projectid + "/" + "depscan.json")
-        blob.upload_from_filename(
-            filedir + "reports/depscan-{}.json".format(project_type)
-        )
-
-    # Remove files from container
-    shutil.rmtree(filedir)
+    # Unset processing
     deleteDB(uid, projectid)
     return Response(status=200)
